+++ ProjectNoName/Network/HTTPMethod.swift
@@ -1,0 +1,13 @@
+//
+//  HTTPMethod.swift
+//  ProjectNoName
+//
+//  Created by Leonir Alves Deolindo on 05/10/20.
+//
+
+import Foundation
+
+enum HTTPMethod: String {
+    case POST
+    case GET
+}
